+++ lib/cmdlang_ast/src/ast.mli
@@ -22,8 +22,8 @@
     to write analysis tools, etc. *)
 
 type 'a or_error_msg = ('a, [ `Msg of string ]) result
-type 'a parse := string -> 'a or_error_msg
-type 'a print := Format.formatter -> 'a -> unit
+type 'a of_string := string -> 'a or_error_msg
+type 'a to_string := 'a -> string
 
 module Nonempty_list : sig
   type 'a t = ( :: ) : 'a * 'a list -> 'a t
@@ -33,8 +33,8 @@ module Param : sig
   type 'a t =
     | Conv :
         { docv : string option
-        ; parse : 'a parse
-        ; print : 'a print
+        ; of_string : 'a of_string
+        ; to_string : 'a to_string
         }
         -> 'a t
     | String : string t